requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
djangorestframework==3.14.0
idna==3.4
lxml==4.9.1
pytz==2022.6
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.12
yapf==0.32.0
